refactor: Integrator's behaviour moves to a mixin, and its population to one place
Third class onto the mixin, and the first whose wrapper did real work in `__init__`.
Splitting that work is what this commit is actually about.

Half of it was already redundant: `__init__` coerced a `noise` dict into a `Noise`, which
LinkML's own `__post_init__` has been doing all along. The other half — populating the
ontology-derived fields — cannot move to a mixin, because `@dataclass` writes `__init__`
and LinkML writes `__post_init__` on the generated class itself, and both win over a base.

It is therefore called explicitly, and from ONE place: whatever integrator an experiment
ends up with is populated in `SimulationExperiment.__init__`. That is stricter than what
it replaces. Before, `Integrator.__init__` populated a directly constructed integrator,
the `__class__` reassignment in `from_datamodel` skipped it and compensated with its own
call, and an integrator loaded from YAML got neither — it was populated only if it also
happened to pass through `from_datamodel`.

Moving it caught that: `test_render_stored_experiment_tvb` began failing with
`'NoneType' object has no attribute 'equation'` because the API path reaches an experiment
without going through `from_datamodel`. It was already relying on population happening in
a constructor that a loaded integrator never ran.

`render_code` and `execute` no longer assign `self.template` and `self.tvb`. Both were
read only inside the method that set them — locals wearing state's clothes — and an
`extra="forbid"` model rejects the assignment outright, so they would have broken the
moment these objects became Pydantic.

`Noise` keeps its wrapper: `@register_pytree_node_class` decorates the class being
defined, which no mixin can supply.

Verified: 2127 passed across the four corpora, package structure, model contract,
experiment API, model runs and the Pydantic loader.
